refactor(backend): move books API handlers from app.js to routes/book.js and controllers/booksController.js
