+++ PBL2/Pustakalaya/backend/requests/request_api.rest
@@ -4,11 +4,11 @@
 ### GET /api/books — with query (expect 200 and JSON with items / totalItems)
 GET {{baseUrl}}/api/books?q=javascript&startIndex=0
 
-### GET /api/books — missing q (expect 400 and error message)
+### GET /api/books — with default query (q = "book")
 GET {{baseUrl}}/api/books/
 
 ### GET /api/books/:id — with valid ID (expect 200 and JSON with book data)
-GET {{baseUrl}}/api/books/U69Hzz2jLPUC
+GET {{baseUrl}}/api/books/jNqFYgLvgLYC
 
 ### GET /api/books/:id — with invalid ID (expect 400 and error message)
 GET {{baseUrl}}/api/books/123
