+++ PBL2/Pustakalaya/backend/routes/book.js
@@ -1,0 +1,9 @@
+const express = require("express");
+const booksController = require("../controllers/booksController");
+
+const router = express.Router();
+
+router.get("/", booksController.searchBooks);
+router.get("/:id", booksController.getBookById);
+
+module.exports = router;
